src/docker_configuration/standard/postgres/requirements.txt
# Postgres dependencies
psycopg2-binary==2.9.9
pandas==2.1.2
sqlalchemy==2.0.23
glob2==0.7
sqlalchemy.orm